app/requirements.txt
aioredis==2.0.1
anyio==3.6.1
async-timeout==4.0.2
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
fastapi==0.82.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
isort==5.10.1
pydantic==1.10.2
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
pika==1.3.0